requirements.txt
langgraph>=0.2.30
openai>=1.40.0
pydantic>=2.7.0
python-dotenv>=1.0.1
google-generativeai>=0.7.2
